Moved button to common component list
